all: fix more warnings

+++ examples/distributed_third_level/combi_example_worker_only.cpp
@@ -156,14 +156,15 @@ int main(int argc, char** argv) {
     // if the file does not exist, one rank creates it
     if (theMPISystem()->getWorldRank() == 0) {
       if (!std::filesystem::exists(interpolationCoordsFile)) {
-        interpolationCoords = montecarlo::getRandomCoordinates(interpolationCoords.size(), dim);
+        interpolationCoords =
+            montecarlo::getRandomCoordinates(static_cast<int>(interpolationCoords.size()), dim);
         h5io::writeValuesToH5File(interpolationCoords, interpolationCoordsFile, "worker_group",
                                   "only");
       }
     }
-#endif // DISCOTEC_USE_HIGHFIVE
-    // get the exact ones e.g. with `wget https://darus.uni-stuttgart.de/api/access/datafile/195524` (1e6)
-    // or `wget https://darus.uni-stuttgart.de/api/access/datafile/195545` (1e5)
+#endif  // DISCOTEC_USE_HIGHFIVE
+    // get the exact ones e.g. with `wget https://darus.uni-stuttgart.de/api/access/datafile/195524`
+    // (1e6) or `wget https://darus.uni-stuttgart.de/api/access/datafile/195545` (1e5)
     interpolationCoords = broadcastParameters::getCoordinatesFromRankZero(
         interpolationCoordsFile, theMPISystem()->getWorldComm());
 
@@ -230,7 +231,7 @@ int main(int argc, char** argv) {
     for (size_t i = 0; i < ncombi; ++i) {
       // run tasks for next time interval
       worker.runAllTasks();
-      auto durationRun = Stats::getDuration("run") / 1000.0;
+      auto durationRun = static_cast<double>(Stats::getDuration("run")) / 1000.0;
       MIDDLE_PROCESS_EXCLUSIVE_SECTION std::cout << getTimeStamp() << "calculation " << i
                                                  << " took: " << durationRun << " seconds"
                                                  << std::endl;
@@ -242,8 +243,8 @@ int main(int argc, char** argv) {
         Stats::stopEvent("write interpolated");
         OTHER_OUTPUT_GROUP_EXCLUSIVE_SECTION {
           MASTER_EXCLUSIVE_SECTION {
-            std::cout << getTimeStamp() << "interpolation " << i
-                      << " took: " << Stats::getDuration("write interpolated") / 1000.0
+            std::cout << getTimeStamp() << "interpolation " << i << " took: "
+                      << static_cast<double>(Stats::getDuration("write interpolated")) / 1000.0
                       << " seconds" << std::endl;
           }
         }
@@ -302,7 +303,7 @@ int main(int argc, char** argv) {
     }
     // run tasks for last time interval
     worker.runAllTasks();
-    auto durationRun = Stats::getDuration("run") / 1000.0;
+    auto durationRun = static_cast<double>(Stats::getDuration("run")) / 1000.0;
     MIDDLE_PROCESS_EXCLUSIVE_SECTION std::cout << getTimeStamp() << "last calculation " << ncombi
                                                << " took: " << durationRun << " seconds"
                                                << std::endl;
@@ -314,9 +315,9 @@ int main(int argc, char** argv) {
       Stats::stopEvent("write interpolated");
       OTHER_OUTPUT_GROUP_EXCLUSIVE_SECTION {
         MASTER_EXCLUSIVE_SECTION {
-          std::cout << getTimeStamp() << "last interpolation " << ncombi
-                    << " took: " << Stats::getDuration("write interpolated") / 1000.0 << " seconds"
-                    << std::endl;
+          std::cout << getTimeStamp() << "last interpolation " << ncombi << " took: "
+                    << static_cast<double>(Stats::getDuration("write interpolated")) / 1000.0
+                    << " seconds" << std::endl;
         }
       }
     }
